requirements.txt
asgiref==3.3.1
Django==3.1.7
django-auth-ldap==2.3.0
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.15.0
python-ldap==3.3.1
pytz==2021.1
sqlparse==0.4.1
